refactor navbars
