--- apps/web/src/components/nav-menu/NavMenu.module.css
@@ -1,313 +1,0 @@
-.navRoot {
-  display: flex;
-  align-items: center;
-  min-width: max-content;
-  gap: var(--space-xs);
-}
-
-@media (--phone) {
-  .desktopMenu {
-    display: none;
-  }
-}
-
-.navList {
-  display: flex;
-  position: relative;
-  flex-wrap: wrap;
-  gap: var(--space-xs);
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.navTrigger,
-.navTriggerActive,
-.accountLink,
-.accountLinkActive {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.375rem;
-  margin: 0;
-  padding: 0.65rem 0.95rem;
-  border: var(--border-width) solid var(--c-border-strong);
-  border-radius: var(--radius-pill);
-  background: var(--c-surface);
-  color: var(--c-text);
-  font: inherit;
-  line-height: 1;
-  text-decoration: none;
-  white-space: nowrap;
-}
-
-.navTrigger,
-.accountLink,
-.accountLinkActive {
-  cursor: pointer;
-}
-
-.navTriggerActive,
-.navTrigger[data-popup-open],
-.navTriggerActive[data-popup-open],
-.accountLinkActive {
-  border-color: var(--c-primary-border);
-  background: var(--c-primary-surface-strong);
-}
-
-.navTrigger:focus-visible,
-.navCard:focus-visible,
-.navCardActive:focus-visible,
-.accountLink:focus-visible,
-.accountLinkActive:focus-visible {
-  outline: 2px solid var(--c-primary-focus);
-  outline-offset: 2px;
-}
-
-@media (hover: hover) and (pointer: fine) {
-  .accountLink:hover,
-  .accountLinkActive:hover {
-    border-color: var(--c-primary-border);
-    background: var(--c-primary-surface-strong);
-  }
-}
-
-.navIcon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: transform var(--duration-fast) ease;
-}
-
-.accountAvatar {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 1.5rem;
-  height: 1.5rem;
-  overflow: hidden;
-  border-radius: 999px;
-  background: var(--c-primary-surface);
-  color: var(--c-text);
-  flex-shrink: 0;
-  margin: -0.25rem 0 -0.25rem -0.25rem;
-}
-
-.accountAvatarImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.accountAvatarFallback {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  font-size: 0.7rem;
-  font-weight: var(--font-weight-semibold);
-  line-height: 1;
-  text-transform: uppercase;
-}
-
-.accountLabel {
-  line-height: 1;
-}
-
-@media (--phone) {
-  .accountLink,
-  .accountLinkActive {
-    width: 2.35rem;
-    height: 2.35rem;
-    padding: 0;
-  }
-
-  .accountAvatar {
-    width: 1.75rem;
-    height: 1.75rem;
-    margin: 0;
-  }
-
-  .accountLabel {
-    display: none;
-  }
-}
-
-.navIcon[data-popup-open] {
-  transform: rotate(180deg);
-}
-
-.navPositioner {
-  --duration: var(--duration-fast);
-  --easing: var(--ease-out);
-  width: var(--positioner-width);
-  height: var(--positioner-height);
-  max-width: min(var(--available-width), calc(100vw - 2rem));
-  transition-property: top, left, right, bottom;
-  transition-duration: var(--duration);
-  transition-timing-function: var(--easing);
-}
-
-.navPositioner[data-instant] {
-  transition: none;
-}
-
-.navPopup {
-  position: relative;
-  width: var(--popup-width);
-  height: var(--popup-height);
-  max-width: min(32rem, calc(100vw - 2rem));
-  border: var(--border-width) solid var(--c-border-popup);
-  border-radius: var(--radius-xs);
-  background: var(--c-surface-solid);
-  color: var(--c-text);
-  box-shadow: var(--shadow-popup);
-  outline: none;
-  transform-origin: var(--transform-origin);
-  transition-property: opacity, transform, width, height;
-  transition-duration: var(--duration);
-  transition-timing-function: var(--easing);
-}
-
-.navPopup[data-starting-style],
-.navPopup[data-ending-style] {
-  opacity: 0;
-  transform: scale(0.96);
-}
-
-.navContent {
-  width: min(32rem, calc(100vw - 2rem));
-  padding: var(--space-sm);
-  box-sizing: border-box;
-  transition:
-    opacity calc(var(--duration) * 0.8) ease,
-    transform var(--duration) var(--easing);
-}
-
-.navContent[data-starting-style],
-.navContent[data-ending-style] {
-  opacity: 0;
-}
-
-.navContent[data-starting-style][data-activation-direction='left'] {
-  transform: translateX(-1rem);
-}
-
-.navContent[data-starting-style][data-activation-direction='right'] {
-  transform: translateX(1rem);
-}
-
-.navContent[data-ending-style][data-activation-direction='left'] {
-  transform: translateX(1rem);
-}
-
-.navContent[data-ending-style][data-activation-direction='right'] {
-  transform: translateX(-1rem);
-}
-
-.navViewport {
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-}
-
-.navLinkList {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(12rem, 1fr));
-  gap: var(--space-xs);
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.navCard,
-.navCardActive {
-  display: block;
-  height: 100%;
-  width: 100%;
-  padding: var(--space-sm);
-  border: var(--border-width) solid transparent;
-  border-radius: var(--radius-xs);
-  background: transparent;
-  color: inherit;
-  font: inherit;
-  text-align: left;
-  text-decoration: none;
-}
-
-.navCard {
-  cursor: pointer;
-}
-
-.navCard:hover,
-.navCard[data-popup-open],
-.navCardActive,
-.navCardActive:hover,
-.navCardActive[data-popup-open] {
-  border-color: var(--c-border-hover);
-  background: var(--c-surface-raised);
-}
-
-.navCardActive {
-  border-color: var(--c-primary-border-soft);
-  background: var(--c-primary-surface);
-}
-
-.navCard:disabled {
-  cursor: wait;
-  opacity: 0.7;
-}
-
-.navCardTitle {
-  margin: 0 0 0.35rem;
-  font-size: 0.95rem;
-  font-weight: var(--font-weight-semibold);
-  line-height: 1.2;
-}
-
-.navCardDescription {
-  margin: 0;
-  color: var(--c-muted-soft);
-  font-size: var(--text-md);
-  line-height: 1.4;
-}
-
-.navArrow {
-  display: block;
-  position: relative;
-  width: 12px;
-  height: 6px;
-  overflow: clip;
-}
-
-.navArrow[data-side='bottom'] {
-  top: -6px;
-}
-
-.navArrow[data-side='top'] {
-  bottom: -6px;
-  rotate: 180deg;
-}
-
-.navArrow[data-side='left'] {
-  right: -9px;
-  rotate: 90deg;
-}
-
-.navArrow[data-side='right'] {
-  left: -9px;
-  rotate: -90deg;
-}
-
-.navArrow::before {
-  content: '';
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  width: calc(6px * sqrt(2));
-  height: calc(6px * sqrt(2));
-  border: var(--border-width) solid var(--c-border-popup);
-  background: var(--c-surface-solid);
-  transform: translate(-50%, 50%) rotate(45deg);
-}
