+++ apps/web/src/components/navbar/Navbar.module.css
@@ -1,0 +1,128 @@
+.navbar {
+  display: flex;
+  align-items: center;
+  min-width: max-content;
+  gap: var(--space-xs);
+}
+
+.trackerList {
+  display: flex;
+  align-items: center;
+  gap: var(--space-xs);
+  margin: 0;
+  padding: 0;
+  list-style: none;
+
+  @media (--phone) {
+    display: none;
+  }
+}
+
+.navLink,
+.navLinkActive,
+.accountLink,
+.accountLinkActive {
+  display: inline-flex;
+  min-height: 2.5rem;
+  align-items: center;
+  justify-content: center;
+  gap: 0.375rem;
+  padding: 0.65rem 0.95rem;
+  border: var(--border-width) solid var(--c-border-strong);
+  border-radius: var(--radius-pill);
+  background: var(--c-surface);
+  color: var(--c-text);
+  line-height: 1;
+  text-decoration: none;
+  white-space: nowrap;
+  transition:
+    background-color var(--duration-fast) ease,
+    border-color var(--duration-fast) ease;
+}
+
+.navLabel {
+  line-height: 1;
+
+  @media (--tablet) {
+    display: none;
+  }
+}
+
+.navLinkActive,
+.accountLinkActive {
+  border-color: var(--c-primary-border);
+  background: var(--c-primary-surface-strong);
+}
+
+.navLink:focus-visible,
+.navLinkActive:focus-visible,
+.accountLink:focus-visible,
+.accountLinkActive:focus-visible {
+  outline: 2px solid var(--c-primary-focus);
+  outline-offset: 2px;
+}
+
+@media (hover: hover) and (pointer: fine) {
+  .navLink:hover,
+  .navLinkActive:hover,
+  .accountLink:hover,
+  .accountLinkActive:hover {
+    border-color: var(--c-primary-border);
+    background: var(--c-primary-surface-strong);
+  }
+}
+
+.accountAvatar {
+  display: inline-flex;
+  width: 1.5rem;
+  height: 1.5rem;
+  flex-shrink: 0;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+  margin: -0.25rem 0 -0.25rem -0.25rem;
+  border-radius: var(--radius-pill);
+  background: var(--c-primary-surface);
+  color: var(--c-text);
+}
+
+.accountAvatarImage {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.accountAvatarFallback {
+  display: flex;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+  font-size: 0.7rem;
+  font-weight: var(--font-weight-semibold);
+  line-height: 1;
+  text-transform: uppercase;
+}
+
+.accountLabel {
+  line-height: 1;
+}
+
+@media (--phone) {
+  .accountLink,
+  .accountLinkActive {
+    width: 2.5rem;
+    height: 2.5rem;
+    padding: 0;
+  }
+
+  .accountAvatar {
+    width: 1.75rem;
+    height: 1.75rem;
+    margin: 0;
+  }
+
+  .accountLabel {
+    display: none;
+  }
+}
